backend/python/django/requirements.txt
aiohttp==3.7.4.post0
anyio==3.5.0
asgiref==3.5.0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
comtypes==1.1.11
cycler==0.11.0
Cython==0.29.27
DateTime==4.3
discord==1.7.3
discord-py-slash-command==3.0.3
discord.py==1.7.3
distlib==0.3.4
Django==4.0.1
dnspython==2.2.0
fastapi==0.73.0
ffmpeg-python==0.2.0
filelock==3.4.2
Flask==2.0.2
Flask-Cors==3.0.10
fonttools==4.29.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
google-api-core==2.5.0
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.55.0
gTTS==2.2.3
h11==0.13.0
httplib2==0.20.4
httptools==0.3.0
idna==3.3
itsdangerous==2.0.1
jaraco.context==4.1.1
Jinja2==3.0.3
Js2Py==0.71
kiwisolver==1.3.2
MarkupSafe==2.1.1
matplotlib==3.5.1
more-itertools==8.12.0
multidict==6.0.2
numpy==1.22.1
oauth2client==4.1.3
opencv-python==4.5.5.64
ordered-set==4.0.2
packaging==21.3
pandas==1.4.1
Pillow==9.0.0
platformdirs==2.5.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
PyDrive==1.3.1
pygooglechart==0.4.0
pyjsparser==2.7.1
pymongo==4.0.1
PyNaCl==1.4.0
pyparsing==3.0.7
pypiwin32==223
Python-Chart==1.39
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pyttsx3==2.90
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
PyYAML==6.0
replit==3.2.4
requests==2.27.1
rsa==4.8
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.2
starlette==0.17.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.4
virtualenv==20.13.1
waiting==1.4.1
watchgod==0.8
websockets==10.2
Werkzeug==2.0.2
wget==3.2
wolframalpha==5.0.0
xmltodict==0.12.0
yarl==1.7.2
Zenora==0.0.3
zope.interface==5.4.0
